folded MuxFilesystem back into CompositeFilesystem
