+++ mrjob/dataproc.py
@@ -35,11 +35,11 @@ import mrjob
 from mrjob.cloud import HadoopInTheCloudJobRunner
 from mrjob.compat import map_version
 from mrjob.conf import combine_dicts
+from mrjob.fs.composite import CompositeFilesystem
 from mrjob.fs.gcs import GCSFilesystem
 from mrjob.fs.gcs import is_gcs_uri
 from mrjob.fs.gcs import parse_gcs_uri
 from mrjob.fs.local import LocalFilesystem
-from mrjob.fs.mux import MuxFilesystem
 from mrjob.logs.counters import _pick_counters
 from mrjob.logs.errors import _format_error
 from mrjob.logs.mixin import LogInterpretationMixin
@@ -403,7 +403,7 @@ class DataprocJobRunner(HadoopInTheCloudJobRunner, LogInterpretationMixin):
         the local filesystem.
         """
         if self._fs is None:
-            self._fs = MuxFilesystem()
+            self._fs = CompositeFilesystem()
 
             self._fs.add_fs('gcs', GCSFilesystem(
                 credentials=self._credentials,
