+++ mrjob/fs/composite.py
@@ -1,6 +1,7 @@
 # Copyright 2009-2012 Yelp and Contributors
 # Copyright 2015 Yelp
 # Copyright 2017 Yelp
+# Copyright 2019 Yelp
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -17,55 +18,123 @@ import logging
 
 from mrjob.fs.base import Filesystem
 
-
 log = logging.getLogger(__name__)
 
 
 class CompositeFilesystem(Filesystem):
-    """Combine multiple filesystem objects to allow access to a variety of
-    storage locations such as the local filesystem, S3, a remote machine via
-    SSH, or HDFS.
-    """
+    """Use one of several filesystems depending on the path/URI.
 
+    This only implements the core :py:class:`~mrjob.fs.base.Filesystem`
+    interface; access extensions by calling the sub-filsystem directly
+    (e.g. ``fs.s3.create_bucket(...)``).
+    """
     def __init__(self, *filesystems):
-        super(CompositeFilesystem, self).__init__()
-        self.filesystems = filesystems
+        # names of sub-filesystems, in the order to call them. (The filesystems
+        # themselves are stored in the attribute with that name.)
+        self._fs_names = []
+
+        # map from fs name to *disable_if* method (see :py:meth:`add`).
+        self._disable_if = {}
+
+        # set of names of filesystems that have been disabled
+        self._disabled = set()
+
+        if filesystems:
+            log.warning('passing filesystems to the constructor is deprecated'
+                        ' and going away in v0.7.0. Use add_fs() instead')
+            for fs in filesystems:
+                # convention is to name filesystems after their module
+                fs_name = fs.__module__.split('.')[-1]
+                self.add_fs(fs, fs_name)
 
     def __getattr__(self, name):
-        # don't confuse pickling when it looks for __getnewargs__, __getstate__
+        # don't confuse pickling (e.g. __getstate__())
         if name.startswith('__'):
             raise AttributeError(name)
 
-        # Forward through to children for backward compatibility
-        for fs in self.filesystems:
+        # go through non-disabled filesystems and pick the first
+        # attribute with a matching name
+        for fs_name in self._fs_names:
+            if fs_name in self._disabled:
+                continue
+
+            fs = getattr(self, fs_name)
             if hasattr(fs, name):
+                log.warning(
+                    'passing %s() through to the top-level filesystem is'
+                    ' deprecated and going away in v0.7.0. Try'
+                    ' fs.%s.%s(...) instead' % (name, fs_name, name))
                 return getattr(fs, name)
+
         raise AttributeError(name)
 
+
+    def add_fs(self, name, fs, disable_if=None):
+        """Add a filesystem.
+
+        :param fs: a :py:class:~mrjob.fs.base.Filesystem to forward calls to.
+        :param name string: Name of this filesystem. It will be directly
+                            accessible through that the attribute with that
+                            name. Recommended usage is the same name as
+                            the module that contains the fs class.
+        :param disable_if: A function called with a single argument, an
+                           exception raised by ``fs``. If it returns true,
+                           futher calls will not be forwarded to ``fs``.
+        """
+        if hasattr(self, name):
+            raise ValueError('name %r is already taken' % name)
+
+        self._fs_names.append(name)
+        setattr(self, name, fs)
+
+        if disable_if:
+            self._disable_if[name] = disable_if
+
     def can_handle_path(self, path):
-        """We can handle a path if any sub-filesystem can."""
-        return any(fs.can_handle_path(path) for fs in self.filesystems)
+        """We can handle any path handled by any (non-disabled) filesystem."""
+        for fs_name in self._fs_names:
+            if fs_name in self._disabled:
+                continue
+
+            fs = getattr(self, fs_name)
+            if fs.can_handle_path(path):
+                return True
+
+        return False
 
     def _do_action(self, action, path, *args, **kwargs):
-        """Call **action** on each filesystem object in turn. If one raises an
-        :py:class:`IOError`, save the exception and try the rest. If none
-        succeed, re-raise the first exception.
-        """
+        """Call method named **action** on the first (non-disabled) filesystem
+        that says it can handle *path*. If it raises an exception, either
+        disable the filesystem and continue, or re-raise the exception."""
+        for fs_name in self._fs_names:
+            if fs_name in self._disabled:
+                continue
 
-        first_exception = None
+            fs = getattr(self, fs_name)
+            if not fs.can_handle_path(path):
+                continue
 
-        for fs in self.filesystems:
-            if fs.can_handle_path(path):
-                try:
-                    return getattr(fs, action)(path, *args, **kwargs)
-                except IOError as e:
-                    if first_exception is None:
-                        first_exception = e
+            try:
+                return getattr(fs, action)(path, *args, **kwargs)
+            except Exception as ex:
+                if (fs_name in self._disable_if and
+                        self._disable_if[fs_name](ex)):
+                    log.debug('disabling %s fs: %r' % (fs_name, ex))
 
-        if first_exception is None:
-            raise IOError("Can't handle path: %s" % path)
-        else:
-            raise first_exception
+                    self._disabled.add(fs_name)
+                else:
+                    raise
+
+        raise IOError("Can't handle path: %s" % path)
+
+    # explicitly implement Filesystem interface. this will come in handy
+
+    def cat(self, path_glob):
+        return self._do_action('cat', path_glob)
+
+    def _cat_file(self, path):
+        # mrjob/runner.py accesses this directly for efficiency
+        return self._do_action('_cat_file', path)
 
     def du(self, path_glob):
         return self._do_action('du', path_glob)
@@ -73,15 +142,11 @@ class CompositeFilesystem(Filesystem):
     def ls(self, path_glob):
         return self._do_action('ls', path_glob)
 
-    def _cat_file(self, path):
-        for line in self._do_action('_cat_file', path):
-            yield line
+    def exists(self, path_glob):
+        return self._do_action('exists', path_glob)
 
     def mkdir(self, path):
         return self._do_action('mkdir', path)
-
-    def exists(self, path_glob):
-        return self._do_action('exists', path_glob)
 
     def join(self, path, *paths):
         return self._do_action('join', path, *paths)
